migrations schema for inventory and sales, and add Livewire components for inventory movements.

+++ database/migrations/tenants/inventory/2025_11_25_211603_add_foreing_key_to_inv_detail_inv_adjusments.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('inv_detail_inv_adjustments', function (Blueprint $table) {
             
-            // 1. Es CRUCIAL cambiar el tipo de dato para que coincida con el id del padre (BigInt Unsigned)
-            // Nota: Si ya tienes datos, asegúrate de que no haya IDs negativos o huérfanos antes de correr esto.
-            $table->unsignedBigInteger('inventoryAdjustmentId')->nullable()->change();
-            $table->unsignedBigInteger('itemId')->nullable()->change();
+            // 1. Cambiar el tipo de dato para que coincida con la tabla padre.
+            // Para 'inv_items.id' es unsignedInteger. Para 'inv_inventory_adjustments.id' (asumo) es BigInteger.
+            $table->unsignedInteger('inventoryAdjustmentId')->nullable()->change();
+            // CORRECCIÓN: Usar unsignedInteger para coincidir con inv_items.id
+            $table->unsignedInteger('itemId')->nullable()->change(); 
             
             // 2. Definir las llaves foráneas
             $table->foreign('inventoryAdjustmentId')
                   ->references('id')
                   ->on('inv_inventory_adjustments')
-                  ->onDelete('cascade'); // Si borras el ajuste, se borran los detalles (opcional: puedes usar 'restrict' o 'set null')
+                  ->onDelete('cascade'); 
 
             $table->foreign('itemId')
                   ->references('id')
@@ -32,24 +30,18 @@ return new class extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('inv_detail_inv_adjustments', function (Blueprint $table) {
             
-            // 1. **ELIMINAR AMBAS LLAVES FORÁNEAS**
-            // Se usa el nombre de la columna para eliminar la FK.
-            // Si Laravel no generó el nombre por defecto (ej: table_column_foreign), 
-            // usar el nombre de columna funciona cuando la definiste con ->foreign('columna').
+            // 1. ELIMINAR AMBAS LLAVES FORÁNEAS
             $table->dropForeign(['inventoryAdjustmentId']);
-            $table->dropForeign(['itemId']); // <--- Faltaba esta
+            $table->dropForeign(['itemId']); 
 
-            // 2. **REVERTIR EL TIPO DE DATO PARA AMBAS COLUMNAS**
-            // Revertir el cambio de tipo de dato (asumiendo que era 'integer' normal antes)
-            $table->integer('inventoryAdjustmentId')->nullable()->change();
-            $table->integer('itemId')->nullable()->change(); // <--- Faltaba esta
+            // 2. REVERTIR EL TIPO DE DATO PARA AMBAS COLUMNAS
+            // Asumiendo que el tipo original era 'integer' sin signo.
+            $table->unsignedInteger('inventoryAdjustmentId')->nullable()->change(); // O el tipo original
+            $table->unsignedInteger('itemId')->nullable()->change(); // O el tipo original
         });
     }
 };
